Save method on Subject class moved to LessonTrackerDbHelper

+++ app/src/main/java/com/example/user/lessontracker/Teachable.java
@@ -1,8 +1,8 @@
 package com.example.user.lessontracker;
 
 public interface Teachable {
-    public int getId();
-    public void setId(int id);
+    public long getId();
+    public void setId(long id);
     public String getTitle();
     public void setTitle(String newTitle);
     public String getDetail();
